fix: created BaseTicket parent class
